+++ services/structure/customer/model.go
@@ -208,6 +208,7 @@ type StructureCustomerModel struct {
 	Domain                       types.String      `tfsdk:"domain"`
 	Email                        types.String      `tfsdk:"email"`
 	GracePeriodDays              types.Int64       `tfsdk:"grace_period_days"`
+	HasActiveHelpdesk            types.Bool        `tfsdk:"has_active_helpdesk"`
 	HasAffiliateLinks            types.Bool        `tfsdk:"has_affiliate_links"`
 	Homepage                     types.String      `tfsdk:"homepage"`
 	HouseNr                      types.String      `tfsdk:"house_nr"`
@@ -311,6 +312,8 @@ func (model *StructureCustomerModel) CopyFrom(ctx context.Context, apiResp Struc
 	model.Email = common.StringPointerValue(apiResp.Email)
 
 	model.GracePeriodDays = types.Int64PointerValue(apiResp.GracePeriodDays)
+
+	model.HasActiveHelpdesk = types.BoolPointerValue(apiResp.HasActiveHelpdesk)
 
 	model.HasAffiliateLinks = types.BoolPointerValue(apiResp.HasAffiliateLinks)